feat(tasks): implement organization-scoped tasks management and filtering
- Add organization context to tasks, allowing users to manage tasks
 specific to their organization.
- Introduce filtering options for tasks based on organization and team.
- Update task creation and updating to associate tasks with an organization
 or team.
- Enhance the UI to support organization-specific task management, including
 dialogs for creating and editing tasks.
- Refactor existing task-related components to accommodate the new
 organization context.
