--- app/(logged-in)/docs/page.tsx
@@ -1,13 +1,0 @@
-export default function DocsPage() {
-  return (
-    <div className="space-y-6">
-      {/* Header */}
-      <div className="space-y-2">
-        <h1 className="text-3xl font-bold tracking-tight">Documentation 🚧</h1>
-        <p className="text-muted-foreground">
-          Coming soon! Documentation is currently under construction. Check back soon!
-        </p>
-      </div>
-    </div>
-  );
-}
