+++ app/(logged-in)/org/[slug]/dashboard/page.tsx
@@ -1,26 +1,759 @@
-/**
- * Organization Dashboard
- * Main dashboard for organization-scoped view
- */
-
 'use client';
 
-import { redirect } from 'next/navigation';
-import { Loader2 } from 'lucide-react';
-import { useActiveOrganization } from '@/hooks/use-active-organization';
+import { Button } from '@/components/ui/button';
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
+import { Badge } from '@/components/ui/badge';
+import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar';
+import { Input } from '@/components/ui/input';
+import { Label } from '@/components/ui/label';
+import { Textarea } from '@/components/ui/textarea';
+import { Checkbox } from '@/components/ui/checkbox';
+import { RadioGroup, RadioGroupItem } from '@/components/ui/radio-group';
+import { Switch } from '@/components/ui/switch';
+import { Slider } from '@/components/ui/slider';
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from '@/components/ui/select';
+import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
+import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert';
+import { Progress } from '@/components/ui/progress';
+import {
+  Table,
+  TableBody,
+  TableCell,
+  TableHead,
+  TableHeader,
+  TableRow,
+} from '@/components/ui/table';
+import {
+  Accordion,
+  AccordionContent,
+  AccordionItem,
+  AccordionTrigger,
+} from '@/components/ui/accordion';
+import { Calendar } from '@/components/ui/calendar';
+import { Skeleton } from '@/components/ui/skeleton';
+import { Toggle } from '@/components/ui/toggle';
+import {
+  AreaChartDemo,
+  StackedAreaChart,
+  LinearAreaChart,
+  GradientAreaChart,
+  StackedBarChart,
+  MixedBarChart,
+} from '@/components/charts/shadcn-charts';
+import { ToggleGroup, ToggleGroupItem } from '@/components/ui/toggle-group';
+import { ScrollArea } from '@/components/ui/scroll-area';
+import {
+  Dialog,
+  DialogContent,
+  DialogDescription,
+  DialogHeader,
+  DialogTitle,
+  DialogTrigger,
+} from '@/components/ui/dialog';
+import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@/components/ui/tooltip';
+import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover';
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuLabel,
+  DropdownMenuSeparator,
+  DropdownMenuTrigger,
+} from '@/components/ui/dropdown-menu';
+import {
+  Sheet,
+  SheetContent,
+  SheetDescription,
+  SheetHeader,
+  SheetTitle,
+  SheetTrigger,
+} from '@/components/ui/sheet';
+import { HoverCard, HoverCardContent, HoverCardTrigger } from '@/components/ui/hover-card';
 
-export default function OrgDashboardPage() {
-  const { isLoading } = useActiveOrganization();
+import {
+  Calendar as CalendarIcon,
+  FileText,
+  BarChart3,
+  Users,
+  MoreHorizontal,
+  Plus,
+  Edit,
+  Trash2,
+  Eye,
+  CheckCircle,
+  AlertCircle,
+  Info,
+  Bold,
+  Italic,
+  Underline,
+  AlignLeft,
+  AlignCenter,
+  AlignRight,
+} from 'lucide-react';
+import { useState } from 'react';
 
-  if (isLoading) {
-    return (
-      <div className="flex items-center justify-center min-h-screen">
-        <Loader2 className="h-8 w-8 animate-spin text-muted-foreground" />
+const users = [
+  {
+    id: 1,
+    name: 'John Doe',
+    email: 'john@example.com',
+    role: 'Admin',
+    status: 'active',
+    avatar: 'https://picsum.photos/seed/user1/32/32',
+  },
+  {
+    id: 2,
+    name: 'Jane Smith',
+    email: 'jane@example.com',
+    role: 'Editor',
+    status: 'active',
+    avatar: 'https://picsum.photos/seed/user2/32/32',
+  },
+  {
+    id: 3,
+    name: 'Bob Johnson',
+    email: 'bob@example.com',
+    role: 'Viewer',
+    status: 'inactive',
+    avatar: 'https://picsum.photos/seed/user3/32/32',
+  },
+];
+
+export default function Page() {
+  const [date, setDate] = useState<Date | undefined>(new Date());
+  const [progress] = useState(65);
+  const [sliderValue, setSliderValue] = useState([50]);
+  const [switchValue, setSwitchValue] = useState(false);
+
+  return (
+    <TooltipProvider>
+      <div className="space-y-8">
+        <div className="flex items-center justify-between">
+          <div className="space-y-2">
+            <h2 className="text-3xl font-bold tracking-tight">Component Showcase</h2>
+            <p className="text-muted-foreground">
+              Explore all available shadcn/ui components with live examples
+            </p>
+          </div>
+        </div>
+
+        {/* Buttons & Actions */}
+        <section className="space-y-4">
+          <h3 className="text-2xl font-semibold">Buttons & Actions</h3>
+          <Card>
+            <CardHeader>
+              <CardTitle>Button Variants</CardTitle>
+              <CardDescription>Different styles and sizes of buttons</CardDescription>
+            </CardHeader>
+            <CardContent className="flex flex-wrap gap-4">
+              <Button>Default</Button>
+              <Button variant="secondary">Secondary</Button>
+              <Button variant="destructive">Destructive</Button>
+              <Button variant="outline">Outline</Button>
+              <Button variant="ghost">Ghost</Button>
+              <Button variant="link">Link</Button>
+              <Button size="sm">Small</Button>
+              <Button size="lg">Large</Button>
+              <Button disabled>Disabled</Button>
+              <Button>
+                <Plus className="mr-2 h-4 w-4" />
+                With Icon
+              </Button>
+            </CardContent>
+          </Card>
+        </section>
+
+        {/* Form Components */}
+        <section className="space-y-4">
+          <h3 className="text-2xl font-semibold">Form Components</h3>
+          <div className="grid gap-4 md:grid-cols-2">
+            <Card>
+              <CardHeader>
+                <CardTitle>Input Fields</CardTitle>
+              </CardHeader>
+              <CardContent className="space-y-4">
+                <div className="space-y-2">
+                  <Label htmlFor="email">Email</Label>
+                  <Input id="email" placeholder="Enter your email" type="email" />
+                </div>
+                <div className="space-y-2">
+                  <Label htmlFor="password">Password</Label>
+                  <Input id="password" placeholder="Enter password" type="password" />
+                </div>
+                <div className="space-y-2">
+                  <Label htmlFor="bio">Bio</Label>
+                  <Textarea id="bio" placeholder="Tell us about yourself" />
+                </div>
+              </CardContent>
+            </Card>
+
+            <Card>
+              <CardHeader>
+                <CardTitle>Selection Controls</CardTitle>
+              </CardHeader>
+              <CardContent className="space-y-4">
+                <div className="space-y-2">
+                  <Label>Framework</Label>
+                  <Select>
+                    <SelectTrigger>
+                      <SelectValue placeholder="Select a framework" />
+                    </SelectTrigger>
+                    <SelectContent>
+                      <SelectItem value="react">React</SelectItem>
+                      <SelectItem value="vue">Vue</SelectItem>
+                      <SelectItem value="angular">Angular</SelectItem>
+                    </SelectContent>
+                  </Select>
+                </div>
+
+                <div className="flex items-center space-x-2">
+                  <Checkbox id="terms" />
+                  <Label htmlFor="terms">Accept terms and conditions</Label>
+                </div>
+
+                <div className="flex items-center space-x-2">
+                  <Switch
+                    id="notifications"
+                    checked={switchValue}
+                    onCheckedChange={setSwitchValue}
+                  />
+                  <Label htmlFor="notifications">Enable notifications</Label>
+                </div>
+
+                <RadioGroup defaultValue="comfortable">
+                  <div className="flex items-center space-x-2">
+                    <RadioGroupItem value="default" id="default" />
+                    <Label htmlFor="default">Default</Label>
+                  </div>
+                  <div className="flex items-center space-x-2">
+                    <RadioGroupItem value="comfortable" id="comfortable" />
+                    <Label htmlFor="comfortable">Comfortable</Label>
+                  </div>
+                </RadioGroup>
+
+                <div className="space-y-2">
+                  <Label>Volume: {sliderValue[0]}</Label>
+                  <Slider
+                    value={sliderValue}
+                    onValueChange={setSliderValue}
+                    max={100}
+                    step={1}
+                    className="w-full"
+                  />
+                </div>
+              </CardContent>
+            </Card>
+          </div>
+        </section>
+
+        {/* Data Display */}
+        <section className="space-y-4">
+          <h3 className="text-2xl font-semibold">Data Display</h3>
+
+          <div className="grid gap-4 md:grid-cols-3">
+            <Card>
+              <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
+                <CardTitle className="text-sm font-medium">Total Users</CardTitle>
+                <Users className="h-4 w-4 text-muted-foreground" />
+              </CardHeader>
+              <CardContent>
+                <div className="text-2xl font-bold">2,350</div>
+                <p className="text-xs text-muted-foreground">+12% from last month</p>
+                <Progress value={progress} className="mt-2" />
+              </CardContent>
+            </Card>
+
+            <Card>
+              <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
+                <CardTitle className="text-sm font-medium">Revenue</CardTitle>
+                <BarChart3 className="h-4 w-4 text-muted-foreground" />
+              </CardHeader>
+              <CardContent>
+                <div className="text-2xl font-bold">$45,231</div>
+                <p className="text-xs text-muted-foreground">+8% from last month</p>
+              </CardContent>
+            </Card>
+
+            <Card>
+              <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
+                <CardTitle className="text-sm font-medium">Active Projects</CardTitle>
+                <FileText className="h-4 w-4 text-muted-foreground" />
+              </CardHeader>
+              <CardContent>
+                <div className="text-2xl font-bold">24</div>
+                <p className="text-xs text-muted-foreground">+3 new this week</p>
+              </CardContent>
+            </Card>
+          </div>
+        </section>
+
+        {/* Badges & Status */}
+        <section className="space-y-4">
+          <h3 className="text-2xl font-semibold">Badges & Status</h3>
+          <Card>
+            <CardHeader>
+              <CardTitle>Badge Variants</CardTitle>
+            </CardHeader>
+            <CardContent className="flex flex-wrap gap-2">
+              <Badge>Default</Badge>
+              <Badge variant="secondary">Secondary</Badge>
+              <Badge variant="destructive">Destructive</Badge>
+              <Badge variant="outline">Outline</Badge>
+              <Badge className="bg-green-500">Success</Badge>
+              <Badge className="bg-yellow-500">Warning</Badge>
+              <Badge className="bg-blue-500">Info</Badge>
+            </CardContent>
+          </Card>
+        </section>
+
+        {/* User Profiles */}
+        <section className="space-y-4">
+          <h3 className="text-2xl font-semibold">User Profiles</h3>
+          <Card>
+            <CardHeader>
+              <CardTitle>Team Members</CardTitle>
+            </CardHeader>
+            <CardContent>
+              <div className="space-y-4">
+                {users.map((user) => (
+                  <div key={user.id} className="flex items-center space-x-4">
+                    <Avatar>
+                      <AvatarImage src={user.avatar} alt={user.name} />
+                      <AvatarFallback className="bg-primary text-primary-foreground">
+                        {user.name
+                          .split(' ')
+                          .map((n) => n[0])
+                          .join('')}
+                      </AvatarFallback>
+                    </Avatar>
+                    <div className="space-y-1">
+                      <p className="text-sm font-medium leading-none">{user.name}</p>
+                      <p className="text-sm text-muted-foreground">{user.email}</p>
+                    </div>
+                    <div className="ml-auto">
+                      <Badge variant={user.status === 'active' ? 'default' : 'secondary'}>
+                        {user.status}
+                      </Badge>
+                    </div>
+                  </div>
+                ))}
+              </div>
+            </CardContent>
+          </Card>
+        </section>
+
+        {/* Alerts & Feedback */}
+        <section className="space-y-4">
+          <h3 className="text-2xl font-semibold">Alerts & Feedback</h3>
+          <div className="space-y-4">
+            <Alert>
+              <Info className="h-4 w-4" />
+              <AlertTitle>Information</AlertTitle>
+              <AlertDescription>
+                This is an informational alert to keep users informed.
+              </AlertDescription>
+            </Alert>
+
+            <Alert className="border-green-200 bg-green-50 text-green-800">
+              <CheckCircle className="h-4 w-4" />
+              <AlertTitle>Success</AlertTitle>
+              <AlertDescription>Your changes have been saved successfully.</AlertDescription>
+            </Alert>
+
+            <Alert variant="destructive">
+              <AlertCircle className="h-4 w-4" />
+              <AlertTitle>Error</AlertTitle>
+              <AlertDescription>Something went wrong. Please try again.</AlertDescription>
+            </Alert>
+          </div>
+        </section>
+
+        {/* Tables */}
+        <section className="space-y-4">
+          <h3 className="text-2xl font-semibold">Data Tables</h3>
+          <Card>
+            <CardHeader>
+              <CardTitle>User Management</CardTitle>
+              <CardDescription>Manage your team members and their permissions</CardDescription>
+            </CardHeader>
+            <CardContent>
+              <Table>
+                <TableHeader>
+                  <TableRow>
+                    <TableHead>Name</TableHead>
+                    <TableHead>Email</TableHead>
+                    <TableHead>Role</TableHead>
+                    <TableHead>Status</TableHead>
+                    <TableHead className="text-right">Actions</TableHead>
+                  </TableRow>
+                </TableHeader>
+                <TableBody>
+                  {users.map((user) => (
+                    <TableRow key={user.id}>
+                      <TableCell className="font-medium">
+                        <div className="flex items-center space-x-2">
+                          <Avatar className="h-8 w-8">
+                            <AvatarImage src={user.avatar} />
+                            <AvatarFallback className="bg-primary text-primary-foreground">
+                              {user.name
+                                .split(' ')
+                                .map((n) => n[0])
+                                .join('')}
+                            </AvatarFallback>
+                          </Avatar>
+                          <span>{user.name}</span>
+                        </div>
+                      </TableCell>
+                      <TableCell>{user.email}</TableCell>
+                      <TableCell>{user.role}</TableCell>
+                      <TableCell>
+                        <Badge variant={user.status === 'active' ? 'default' : 'secondary'}>
+                          {user.status}
+                        </Badge>
+                      </TableCell>
+                      <TableCell className="text-right">
+                        <DropdownMenu>
+                          <DropdownMenuTrigger asChild>
+                            <Button variant="ghost" className="h-8 w-8 p-0">
+                              <MoreHorizontal className="h-4 w-4" />
+                            </Button>
+                          </DropdownMenuTrigger>
+                          <DropdownMenuContent align="end">
+                            <DropdownMenuLabel>Actions</DropdownMenuLabel>
+                            <DropdownMenuItem>
+                              <Edit className="mr-2 h-4 w-4" />
+                              Edit
+                            </DropdownMenuItem>
+                            <DropdownMenuItem>
+                              <Eye className="mr-2 h-4 w-4" />
+                              View details
+                            </DropdownMenuItem>
+                            <DropdownMenuSeparator />
+                            <DropdownMenuItem className="text-destructive">
+                              <Trash2 className="mr-2 h-4 w-4" />
+                              Delete
+                            </DropdownMenuItem>
+                          </DropdownMenuContent>
+                        </DropdownMenu>
+                      </TableCell>
+                    </TableRow>
+                  ))}
+                </TableBody>
+              </Table>
+            </CardContent>
+          </Card>
+        </section>
+
+        {/* Layout Components */}
+        <section className="space-y-4">
+          <h3 className="text-2xl font-semibold">Layout & Navigation</h3>
+
+          <div className="grid gap-4 md:grid-cols-2">
+            <Card>
+              <CardHeader>
+                <CardTitle>Tabs</CardTitle>
+              </CardHeader>
+              <CardContent>
+                <Tabs defaultValue="overview" className="w-full">
+                  <TabsList className="grid w-full grid-cols-3">
+                    <TabsTrigger value="overview">Overview</TabsTrigger>
+                    <TabsTrigger value="analytics">Analytics</TabsTrigger>
+                    <TabsTrigger value="reports">Reports</TabsTrigger>
+                  </TabsList>
+                  <TabsContent value="overview" className="space-y-2">
+                    <p className="text-sm text-muted-foreground">
+                      Overview of your dashboard metrics and recent activity.
+                    </p>
+                  </TabsContent>
+                  <TabsContent value="analytics" className="space-y-2">
+                    <p className="text-sm text-muted-foreground">
+                      Detailed analytics and performance insights.
+                    </p>
+                  </TabsContent>
+                  <TabsContent value="reports" className="space-y-2">
+                    <p className="text-sm text-muted-foreground">
+                      Generate and download custom reports.
+                    </p>
+                  </TabsContent>
+                </Tabs>
+              </CardContent>
+            </Card>
+
+            <Card>
+              <CardHeader>
+                <CardTitle>Accordion</CardTitle>
+              </CardHeader>
+              <CardContent>
+                <Accordion type="single" collapsible className="w-full">
+                  <AccordionItem value="item-1">
+                    <AccordionTrigger>Getting Started</AccordionTrigger>
+                    <AccordionContent>
+                      Learn the basics of using our platform with this comprehensive guide.
+                    </AccordionContent>
+                  </AccordionItem>
+                  <AccordionItem value="item-2">
+                    <AccordionTrigger>Advanced Features</AccordionTrigger>
+                    <AccordionContent>
+                      Explore advanced features and customization options.
+                    </AccordionContent>
+                  </AccordionItem>
+                  <AccordionItem value="item-3">
+                    <AccordionTrigger>API Documentation</AccordionTrigger>
+                    <AccordionContent>
+                      Complete API reference and integration examples.
+                    </AccordionContent>
+                  </AccordionItem>
+                </Accordion>
+              </CardContent>
+            </Card>
+          </div>
+        </section>
+
+        {/* Interactive Components */}
+        <section className="space-y-4">
+          <h3 className="text-2xl font-semibold">Interactive Components</h3>
+
+          <div className="grid gap-4 md:grid-cols-2 lg:grid-cols-3">
+            <Card>
+              <CardHeader>
+                <CardTitle>Dialog</CardTitle>
+              </CardHeader>
+              <CardContent>
+                <Dialog>
+                  <DialogTrigger asChild>
+                    <Button>Open Dialog</Button>
+                  </DialogTrigger>
+                  <DialogContent>
+                    <DialogHeader>
+                      <DialogTitle>Edit Profile</DialogTitle>
+                      <DialogDescription>
+                        Make changes to your profile here. Click save when you&apos;re done.
+                      </DialogDescription>
+                    </DialogHeader>
+                    <div className="space-y-4">
+                      <div>
+                        <Label htmlFor="name">Name</Label>
+                        <Input id="name" defaultValue="John Doe" />
+                      </div>
+                      <div>
+                        <Label htmlFor="email">Email</Label>
+                        <Input id="email" defaultValue="john@example.com" />
+                      </div>
+                    </div>
+                  </DialogContent>
+                </Dialog>
+              </CardContent>
+            </Card>
+
+            <Card>
+              <CardHeader>
+                <CardTitle>Sheet</CardTitle>
+              </CardHeader>
+              <CardContent>
+                <Sheet>
+                  <SheetTrigger asChild>
+                    <Button variant="outline">Open Sheet</Button>
+                  </SheetTrigger>
+                  <SheetContent>
+                    <SheetHeader>
+                      <SheetTitle>Settings</SheetTitle>
+                      <SheetDescription>Configure your application settings.</SheetDescription>
+                    </SheetHeader>
+                    <div className="space-y-4 mt-4">
+                      <div className="flex items-center space-x-2">
+                        <Switch id="dark-mode" />
+                        <Label htmlFor="dark-mode">Dark mode</Label>
+                      </div>
+                      <div className="flex items-center space-x-2">
+                        <Switch id="notifications" />
+                        <Label htmlFor="notifications">Notifications</Label>
+                      </div>
+                    </div>
+                  </SheetContent>
+                </Sheet>
+              </CardContent>
+            </Card>
+
+            <Card>
+              <CardHeader>
+                <CardTitle>Popover</CardTitle>
+              </CardHeader>
+              <CardContent>
+                <Popover>
+                  <PopoverTrigger asChild>
+                    <Button variant="outline">
+                      <CalendarIcon className="mr-2 h-4 w-4" />
+                      Pick a date
+                    </Button>
+                  </PopoverTrigger>
+                  <PopoverContent className="w-auto p-0" align="start">
+                    <Calendar mode="single" selected={date} onSelect={setDate} initialFocus />
+                  </PopoverContent>
+                </Popover>
+              </CardContent>
+            </Card>
+          </div>
+        </section>
+
+        {/* Tooltips & Hover Cards */}
+        <section className="space-y-4">
+          <h3 className="text-2xl font-semibold">Tooltips & Hover Cards</h3>
+          <Card>
+            <CardHeader>
+              <CardTitle>Interactive Help</CardTitle>
+            </CardHeader>
+            <CardContent className="flex gap-4">
+              <Tooltip>
+                <TooltipTrigger asChild>
+                  <Button variant="outline">Hover me</Button>
+                </TooltipTrigger>
+                <TooltipContent>
+                  <p>This is a helpful tooltip</p>
+                </TooltipContent>
+              </Tooltip>
+
+              <HoverCard>
+                <HoverCardTrigger asChild>
+                  <Button variant="link">@johndev</Button>
+                </HoverCardTrigger>
+                <HoverCardContent className="w-80">
+                  <div className="flex justify-between space-x-4">
+                    <Avatar>
+                      <AvatarImage src="https://picsum.photos/seed/hover1/40/40" />
+                      <AvatarFallback className="bg-primary text-primary-foreground">
+                        JD
+                      </AvatarFallback>
+                    </Avatar>
+                    <div className="space-y-1">
+                      <h4 className="text-sm font-semibold">John Developer</h4>
+                      <p className="text-sm">
+                        Senior Frontend Developer with 5 years of experience in React.
+                      </p>
+                      <div className="flex items-center pt-2">
+                        <CalendarIcon className="mr-2 h-4 w-4 opacity-70" />
+                        <span className="text-xs text-muted-foreground">Joined December 2021</span>
+                      </div>
+                    </div>
+                  </div>
+                </HoverCardContent>
+              </HoverCard>
+            </CardContent>
+          </Card>
+        </section>
+
+        {/* Toggle & Formatting */}
+        <section className="space-y-4">
+          <h3 className="text-2xl font-semibold">Toggle & Formatting</h3>
+          <Card>
+            <CardHeader>
+              <CardTitle>Text Formatting</CardTitle>
+            </CardHeader>
+            <CardContent className="space-y-4">
+              <div className="flex gap-2">
+                <Toggle>
+                  <Bold className="h-4 w-4" />
+                </Toggle>
+                <Toggle>
+                  <Italic className="h-4 w-4" />
+                </Toggle>
+                <Toggle>
+                  <Underline className="h-4 w-4" />
+                </Toggle>
+              </div>
+
+              <ToggleGroup type="single" defaultValue="left">
+                <ToggleGroupItem value="left">
+                  <AlignLeft className="h-4 w-4" />
+                </ToggleGroupItem>
+                <ToggleGroupItem value="center">
+                  <AlignCenter className="h-4 w-4" />
+                </ToggleGroupItem>
+                <ToggleGroupItem value="right">
+                  <AlignRight className="h-4 w-4" />
+                </ToggleGroupItem>
+              </ToggleGroup>
+            </CardContent>
+          </Card>
+        </section>
+
+        {/* Loading States */}
+        <section className="space-y-4">
+          <h3 className="text-2xl font-semibold">Loading States</h3>
+          <Card>
+            <CardHeader>
+              <CardTitle>Skeleton Loading</CardTitle>
+            </CardHeader>
+            <CardContent className="space-y-4">
+              <div className="flex items-center space-x-4">
+                <Skeleton className="h-12 w-12 rounded-full" />
+                <div className="space-y-2">
+                  <Skeleton className="h-4 w-[250px]" />
+                  <Skeleton className="h-4 w-[200px]" />
+                </div>
+              </div>
+              <Skeleton className="h-[200px] w-full" />
+            </CardContent>
+          </Card>
+        </section>
+
+        {/* Charts */}
+        <section className="space-y-4">
+          <h3 className="text-2xl font-semibold">Charts</h3>
+          <p className="text-muted-foreground">
+            Beautiful charts built with shadcn/ui and Recharts. Copy and paste into your apps.
+          </p>
+
+          {/* First Row - Area Charts */}
+          <div className="grid gap-6 md:grid-cols-2 lg:grid-cols-3">
+            <AreaChartDemo />
+            <StackedAreaChart />
+            <LinearAreaChart />
+          </div>
+
+          {/* Second Row - Bar Charts and Gradient */}
+          <div className="grid gap-6 md:grid-cols-2 lg:grid-cols-3">
+            <StackedBarChart />
+            <MixedBarChart />
+            <GradientAreaChart />
+          </div>
+        </section>
+
+        {/* Scroll Area */}
+        <section className="space-y-4">
+          <h3 className="text-2xl font-semibold">Scroll Area</h3>
+          <Card>
+            <CardHeader>
+              <CardTitle>Recent Activity</CardTitle>
+            </CardHeader>
+            <CardContent>
+              <ScrollArea className="h-[200px] w-full">
+                <div className="space-y-4">
+                  {Array.from({ length: 10 }).map((_, i) => (
+                    <div key={i} className="flex items-center space-x-4">
+                      <Avatar className="h-8 w-8">
+                        <AvatarImage src={`https://picsum.photos/seed/activity${i}/32/32`} />
+                        <AvatarFallback className="bg-primary text-primary-foreground">
+                          U{i}
+                        </AvatarFallback>
+                      </Avatar>
+                      <div className="space-y-1">
+                        <p className="text-sm font-medium">User {i + 1} performed an action</p>
+                        <p className="text-xs text-muted-foreground">{i + 1} minutes ago</p>
+                      </div>
+                    </div>
+                  ))}
+                </div>
+              </ScrollArea>
+            </CardContent>
+          </Card>
+        </section>
       </div>
-    );
-  }
-
-  // Redirect to the non-org dashboard for now
-  // In the future, this would show org-specific dashboard
-  redirect('/dashboard');
+    </TooltipProvider>
+  );
 }
