+++ lib/trpc/routers/tasks.ts
@@ -3,7 +3,7 @@
  * Handles CRUD operations for the todo list with server-side filtering
  */
 
-import { eq, and, desc, or, like } from 'drizzle-orm';
+import { eq, and, desc, or, like, isNull } from 'drizzle-orm';
 import { db } from '@/lib/db/drizzle';
 import { tasks } from '@/lib/db/schema';
 import { router, protectedProcedure } from '../init';
@@ -19,9 +19,30 @@ import {
 export const tasksRouter = router({
   /**
    * Get all tasks for the authenticated user with server-side filtering
+   * Supports both personal tasks and org-scoped tasks
    */
   list: protectedProcedure.input(taskListFiltersSchema).query(async ({ ctx, input }) => {
     const conditions = [eq(tasks.clerkUserId, ctx.userId)];
+
+    // Filter by organization (or personal tasks if null)
+    if (input?.organizationId !== undefined) {
+      if (input.organizationId === null) {
+        // Personal tasks only (no org)
+        conditions.push(isNull(tasks.organizationId));
+      } else {
+        // Org-specific tasks
+        conditions.push(eq(tasks.organizationId, input.organizationId));
+      }
+    }
+
+    // Filter by team
+    if (input?.teamId !== undefined) {
+      if (input.teamId === null) {
+        conditions.push(isNull(tasks.teamId));
+      } else {
+        conditions.push(eq(tasks.teamId, input.teamId));
+      }
+    }
 
     // Filter by completion status
     if (input?.completed !== undefined) {
@@ -61,13 +82,15 @@ export const tasksRouter = router({
   }),
 
   /**
-   * Create a new task
+   * Create a new task (supports personal and org-scoped tasks)
    */
   create: protectedProcedure.input(createTaskSchema).mutation(async ({ ctx, input }) => {
     const [task] = await db
       .insert(tasks)
       .values({
         clerkUserId: ctx.userId,
+        organizationId: input.organizationId ?? null,
+        teamId: input.teamId ?? null,
         title: input.title,
         description: input.description ?? null,
         priority: input.priority,
@@ -115,6 +138,8 @@ export const tasksRouter = router({
     if (input.completed !== undefined) updateData.completed = input.completed ? 'true' : 'false';
     if (input.priority !== undefined) updateData.priority = input.priority;
     if (input.dueDate !== undefined) updateData.dueDate = input.dueDate;
+    if (input.organizationId !== undefined) updateData.organizationId = input.organizationId;
+    if (input.teamId !== undefined) updateData.teamId = input.teamId;
 
     const [updatedTask] = await db
       .update(tasks)
